refactor(conditions): move conditions to their respective modules
