+++ examples/taco/nodejs/src/index.ts
@@ -46,7 +46,7 @@ const encryptToBytes = async (messageString: string) => {
   const message = toBytes(messageString);
   console.log(format('Encrypting message ("%s") ...', messageString));
 
-  const hasPositiveBalance = new conditions.base.RpcCondition({
+  const hasPositiveBalance = new conditions.base.rpc.RpcCondition({
     chain: 80001,
     method: 'eth_getBalance',
     parameters: [':userAddress', 'latest'],
