+++ examples/taco/react/src/App.tsx
@@ -65,7 +65,7 @@ function App() {
     setEncrypting(true)
     try {
       const signer = provider.getSigner();
-      const hasPositiveBalance = new conditions.base.RpcCondition({
+      const hasPositiveBalance = new conditions.base.rpc.RpcCondition({
         chain: 80001,
         method: 'eth_getBalance',
         parameters: [':userAddress', 'latest'],
