+++ packages/taco/src/conditions/index.ts
@@ -4,5 +4,6 @@ import * as predefined from './predefined';
 export * as compound from './compound-condition';
 export * as condition from './condition';
 export * as conditionExpr from './condition-expr';
+export { ConditionFactory } from './condition-factory';
 export * as context from './context';
 export { base, predefined };
