+++ packages/taco/src/conditions/predefined/erc721.ts
@@ -2,7 +2,7 @@ import {
   ContractCondition,
   ContractConditionProps,
   ContractConditionType,
-} from '../base';
+} from '../base/contract';
 import { USER_ADDRESS_PARAM } from '../const';
 
 type ERC721OwnershipFields = 'contractAddress' | 'chain' | 'parameters';
