+++ packages/taco/test/conditions/base/condition.test.ts
@@ -1,7 +1,7 @@
 import { TEST_CONTRACT_ADDR } from '@nucypher/test-utils';
 import { describe, expect, it } from 'vitest';
 
-import { ContractCondition } from '../../../src/conditions/base';
+import { ContractCondition } from '../../../src/conditions/base/contract';
 import { Condition } from '../../../src/conditions/condition';
 import { ERC721Ownership } from '../../../src/conditions/predefined/erc721';
 import { fakeCondition, testContractConditionObj } from '../../test-utils';
