+++ packages/taco/test/conditions/condition-expr.test.ts
@@ -5,14 +5,15 @@ import { SemVer } from 'semver';
 import { beforeAll, describe, expect, it } from 'vitest';
 
 import {
-  CompoundCondition,
   ContractCondition,
   ContractConditionProps,
-  RpcCondition,
-  RpcConditionType,
+} from '../../src/conditions/base/contract';
+import { RpcCondition, RpcConditionType } from '../../src/conditions/base/rpc';
+import {
   TimeCondition,
   TimeConditionProps,
-} from '../../src/conditions/base';
+} from '../../src/conditions/base/time';
+import { CompoundCondition } from '../../src/conditions/compound-condition';
 import { ConditionExpression } from '../../src/conditions/condition-expr';
 import { USER_ADDRESS_PARAM } from '../../src/conditions/const';
 import { ERC721Balance } from '../../src/conditions/predefined/erc721';
