+++ packages/taco/test/conditions/conditions.test.ts
@@ -3,7 +3,7 @@ import { fakeProvider, fakeSigner } from '@nucypher/test-utils';
 import { beforeAll, describe, expect, it } from 'vitest';
 
 import { initialize } from '../../src';
-import { CompoundCondition } from '../../src/conditions/base';
+import { CompoundCondition } from '../../src/conditions/compound-condition';
 import { SUPPORTED_CHAIN_IDS } from '../../src/conditions/const';
 import { ConditionContext } from '../../src/conditions/context';
