+++ packages/taco/test/test-utils.ts
@@ -38,15 +38,19 @@ import {
   ContractConditionProps,
   ContractConditionType,
   FunctionAbiProps,
-  ReturnValueTestProps,
+} from '../src/conditions/base/contract';
+import {
   RpcConditionProps,
   RpcConditionType,
+} from '../src/conditions/base/rpc';
+import {
+  TimeConditionMethod,
   TimeConditionProps,
   TimeConditionType,
-} from '../src/conditions/base';
-import { TimeConditionMethod } from '../src/conditions/base/time';
+} from '../src/conditions/base/time';
 import { ConditionExpression } from '../src/conditions/condition-expr';
 import { ERC721Balance } from '../src/conditions/predefined/erc721';
+import { ReturnValueTestProps } from '../src/conditions/shared';
 import { DkgClient, DkgRitual } from '../src/dkg';
 import { encryptMessage } from '../src/tdec';
